# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DataForSEO Intelligence API
  version: 1.0.0
  summary: Multi-tenant SEO research API on top of DataForSEO v3.
  description: 'A multi-tenant SaaS SEO platform. Every billed call flows through a three-tier cost engine
    (Redis -> Postgres -> upstream) with singleflight coalescing and stale-while-revalidate, so repeat
    reads are $0.


    **Conventions**

    - All endpoints are versioned under `/api/v1`.

    - Every data response embeds a `meta` object (`from_cache`, `cost_cents`, `source`, `latency_ms`).

    - All money values are integer **USD cents**.

    - Auth is JWT Bearer; access tokens are short-lived, refreshed via `/auth/refresh`.

    - Per-organization monthly quota is enforced before any billed call (HTTP 402).'
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
paths:
...
  /api/v1/keywords/volume:
    post:
      tags:
      - keywords
      summary: Volume
      operationId: volume_api_v1_keywords_volume_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolumeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolumeResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/keywords/trends:
    post:
      tags:
      - keywords
      summary: Trends
      operationId: trends_api_v1_keywords_trends_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrendsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/keywords/suggestions:
    post:
      tags:
      - keywords
      summary: Suggestions
      operationId: suggestions_api_v1_keywords_suggestions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeedRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeywordListResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/keywords/related:
    post:
      tags:
      - keywords
      summary: Related
      operationId: related_api_v1_keywords_related_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeedRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeywordListResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/keywords/ideas:
    post:
      tags:
      - keywords
      summary: Ideas
      operationId: ideas_api_v1_keywords_ideas_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdeasRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeywordListResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/keywords/paa:
    post:
      tags:
      - keywords
      summary: Paa
      operationId: paa_api_v1_keywords_paa_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaaRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaaResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    IdeasRequest:
      properties:
        keywords:
          items:
            type: string
          type: array
          maxItems: 200
          minItems: 1
          title: Keywords
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
        limit:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          default: 100
      type: object
      required:
      - keywords
      title: IdeasRequest
...
    KeywordListResponse:
      properties:
        rows:
          items:
            $ref: '#/components/schemas/KeywordRow'
          type: array
          title: Rows
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - rows
      - meta
      title: KeywordListResponse
    KeywordRow:
      properties:
        keyword:
          type: string
          title: Keyword
        search_volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Search Volume
        cpc:
          anyOf:
          - type: number
          - type: 'null'
          title: Cpc
        competition:
          anyOf:
          - type: number
          - type: 'null'
          title: Competition
        keyword_difficulty:
          anyOf:
          - type: integer
          - type: 'null'
          title: Keyword Difficulty
        intent:
          anyOf:
          - type: string
          - type: 'null'
          title: Intent
      type: object
      required:
      - keyword
      title: KeywordRow
...
    Meta:
      properties:
        from_cache:
          type: boolean
          title: From Cache
        cost_cents:
          type: integer
          title: Cost Cents
        source:
          type: string
          title: Source
        latency_ms:
          type: integer
          title: Latency Ms
      type: object
      required:
      - from_cache
      - cost_cents
      - source
      - latency_ms
      title: Meta
    MonthlyPoint:
      properties:
        year:
          anyOf:
          - type: integer
          - type: 'null'
          title: Year
        month:
          anyOf:
          - type: integer
          - type: 'null'
          title: Month
        volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Volume
      type: object
      title: MonthlyPoint
...
    PaaItem:
      properties:
        question:
          type: string
          title: Question
          default: ''
        answer:
          anyOf:
          - type: string
          - type: 'null'
          title: Answer
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
      type: object
      title: PaaItem
    PaaRequest:
      properties:
        keyword:
          type: string
          maxLength: 300
          minLength: 1
          title: Keyword
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
      type: object
      required:
      - keyword
      title: PaaRequest
    PaaResponse:
      properties:
        keyword:
          type: string
          title: Keyword
        paa:
          items:
            $ref: '#/components/schemas/PaaItem'
          type: array
          title: Paa
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - keyword
      - paa
      - meta
      title: PaaResponse
...
    SeedRequest:
      properties:
        seed:
          type: string
          maxLength: 300
          minLength: 1
          title: Seed
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
        limit:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          default: 100
      type: object
      required:
      - seed
      title: SeedRequest
...
    TrendsPoint:
      properties:
        date:
          anyOf:
          - type: string
          - type: 'null'
          title: Date
        values:
          items:
            anyOf:
            - type: integer
            - type: 'null'
          type: array
          title: Values
          default: []
      type: object
      title: TrendsPoint
    TrendsRequest:
      properties:
        keywords:
          items:
            type: string
          type: array
          maxItems: 5
          minItems: 1
          title: Keywords
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
        time_range:
          type: string
          title: Time Range
          default: past_12_months
        date_from:
          anyOf:
          - type: string
            maxLength: 10
          - type: 'null'
          title: Date From
        date_to:
          anyOf:
          - type: string
            maxLength: 10
          - type: 'null'
          title: Date To
      type: object
      required:
      - keywords
      title: TrendsRequest
    TrendsResponse:
      properties:
        keywords:
          items:
            type: string
          type: array
          title: Keywords
        graph:
          items:
            $ref: '#/components/schemas/TrendsPoint'
          type: array
          title: Graph
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - keywords
      - graph
      - meta
      title: TrendsResponse
...
    VolumeRequest:
      properties:
        keywords:
          items:
            type: string
          type: array
          maxItems: 1000
          minItems: 1
          title: Keywords
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
      type: object
      required:
      - keywords
      title: VolumeRequest
    VolumeResponse:
      properties:
        rows:
          items:
            $ref: '#/components/schemas/VolumeRow'
          type: array
          title: Rows
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - rows
      - meta
      title: VolumeResponse
    VolumeRow:
      properties:
        keyword:
          type: string
          title: Keyword
        search_volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Search Volume
        cpc:
          anyOf:
          - type: number
          - type: 'null'
          title: Cpc
        competition:
          anyOf:
          - type: integer
          - type: 'null'
          title: Competition
        competition_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Competition Level
        monthly_searches:
          items:
            $ref: '#/components/schemas/MonthlyPoint'
          type: array
          title: Monthly Searches
          default: []
      type: object
      required:
      - keyword
      title: VolumeRow
...
    Problem:
      type: object
      description: RFC 7807 Problem Details (media type `application/problem+json`).
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
          format: uri
          examples:
          - https://docs.seo-intelligence.app/errors/quota-exceeded
        title:
          type: string
          examples:
          - Payment Required
        status:
          type: integer
          examples:
          - 402
        detail:
          type: string
          examples:
          - Monthly organization quota exhausted.
        instance:
          type: string
          examples:
          - /api/v1/serp/ranking
        errors:
          type: array
          description: Field-level validation failures (422 only).
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from `/auth/login` or `/auth/refresh`.
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PaymentRequired:
      description: Monthly organization quota exhausted — top up or wait for the next cycle.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Upstream rate limit hit; retry after the indicated delay.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
    ValidationFailed:
      description: Request body or parameters failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.example.com
  description: Production
...
security:
- BearerAuth: []